string: Support targets with 16-bit ints in optimized code
The optimized string code is only used for 32-bit aligned data, but
the alignment checks were using 'long' instead of 'uintptr_t', causing
problems on targets with 16-bit pointers. Also, some computations need
explicit uint32_t casts to avoid truncation under the standard integer
promotion rules.

+++ newlib/libc/string/memcmp.c
@@ -46,11 +46,11 @@ QUICKREF
 */
 
 #include <string.h>
-
+#include <stdint.h>
 
 /* Nonzero if either X or Y is not aligned on a "long" boundary.  */
 #define UNALIGNED(X, Y) \
-  (((long)X & (sizeof (long) - 1)) | ((long)Y & (sizeof (long) - 1)))
+  (((uintptr_t)X & (sizeof (long) - 1)) | ((uintptr_t)Y & (sizeof (long) - 1)))
 
 /* How many bytes are copied each iteration of the word copy loop.  */
 #define LBLOCKSIZE (sizeof (long))
